+++ src/ui.rs
@@ -326,7 +326,6 @@ impl Ui {
     pub fn run(self) {
         let options = eframe::NativeOptions {
             always_on_top: true,
-            decorated: false,
             initial_window_size: Some(egui::Vec2::new(180.0, 128.0)),
             resizable: false,
             visible: false,